requirements.txt
alembic==1.7.7
anyio==3.6.2
asgiref==3.4.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
contextlib2==21.6.0
contextvars==2.4
cryptography==40.0.2
dataclasses==0.8
dnspython==2.2.1
docopt==0.6.2
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.83.0
greenlet==2.0.2
h11==0.13.0
httptools==0.3.0
idna==3.6
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Js2Py==0.74
Mako==1.1.6
MarkupSafe==2.0.1
orjson==3.6.1
packaging==21.3
passlib==1.7.4
psycopg2 @ file:///C:/Users/karan/pipwin/psycopg2-2.8.6-cp36-cp36m-win_amd64.whl
pyasn1==0.5.1
pycparser==2.21
pydantic==1.9.2
pyjsparser==2.7.1
pyparsing==3.0.7
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
requests==2.27.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.23
starlette==0.19.1
typing_extensions==4.1.1
tzdata==2023.4
tzlocal==4.2
ujson==4.0.1
urllib3==1.26.18
uvicorn==0.16.0
watchgod==0.7
websockets==9.1
zipp==3.6.0
